More refactoring, and changed incoming message for Activity API
