+++ libraries/botbuilder-adapter-twitter/src/index.ts
@@ -4,4 +4,5 @@
 
 export * from './schema';
 export * from './twitterAdapter';
+export * from './twitterWebhook';
 export * from './util';
